+++ CHANGELOG.md
@@ -4,13 +4,15 @@ with `zswap` being tracked in [Changelog Zswap](./CHANGELOG_zswap.md).
 
 # Change Log
 
-## Unreleased (9.0)
+## Ledger 9.0.1.0-alpha.1
 
 - feat: add explicit price floor, denominated in full blocks, and governed by
   ledger parameters.
 - feat: add support for ECDSA signatures
 - feat: add `UnlockToTreasury` system transaction, moving funds from the locked
   pool to the treasury.
+- feat: expose the `contractLog` variant of `EventDetails` in the ledger-wasm
+  matching the new on-chain event format.
 - fix: correctly exclude the identity point during coin ciphertext decryption
 
 ## Unreleased (8.2)